make Irrlicht compile with VS 6.0 and add project file for that IDE
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@105 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CFileList.cpp
@@ -40,7 +40,7 @@ CFileList::CFileList()
 	Path = tmp;
 
 	struct _finddata_t c_file;
-	intptr_t hFile;
+	long hFile;
 	FileEntry entry;
 
 	if( (hFile = _findfirst( "*", &c_file )) != -1L )
